requirements.txt
Flask==2.3.2
Pillow==9.5.0
Pillow==10.0.0
torch==2.0.1
transformers==4.29.1
ultralytics==8.0.99
